Updated DataStorageAPIImpl and added ProductSumsCalculatorAPIImpl
added reference to EngineProcessAPI and took any methods that are related to calulcating the product-sum and moved them into a new file called ProductSumsCalculatorAPIImpl
