+++ src/productsums/impl/process/DataStorageProcessAPIImpl.java
@@ -3,33 +3,38 @@ package productsums.impl.process;
 import productsums.api.process.DataStorageProcessAPI;
 import productsums.models.process.DataStorageProcessRequest;
 import productsums.models.process.DataStorageProcessResponse;
+import productsums.api.compute.EngineProcessAPI;
+import productsums.api.compute.EngineProcessPrototype;
+import productsums.models.compute.EngineInput;
+import productsums.models.compute.EngineOutput;
 import productsums.utils.FileReaderUtil;
 import productsums.utils.FileWriterUtil;
 
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
-import java.util.Set;
 
-/**
- * Implementation of the DataStorageProcessAPI that calculates minimal product-sum numbers.
- * A number N is called a product-sum number if it can be expressed as both:
- * 1. A sum of k numbers: N = a1 + a2 + ... + ak
- * 2. A product of the same k numbers: N = a1 × a2 × ... × ak
- * This implementation finds the minimal such N for each k in a given range.
- */
 public class DataStorageProcessAPIImpl implements DataStorageProcessAPI {
 
+    // Reference to the EngineProcessAPI
+    private EngineProcessAPI engineProcessAPI;
+
+    public DataStorageProcessAPIImpl() {
+        // Instantiate the EngineProcessAPI (could be injected instead)
+        this.engineProcessAPI = new EngineProcessPrototype();
+    }
+
     /**
-     * 1. Validates the input request
-     * 2. Reads input data if source is provided
-     * 3. Computes product-sum numbers
-     * 4. Writes results if destination is provided
+     * Processes the data by:
+     * 1. Validating the request.
+     * 2. Optionally reading input from a file.
+     * 3. Delegating computation of product-sum numbers to the EngineProcessAPI.
+     * 4. Optionally writing results to an output file.
      */
     @Override
     public DataStorageProcessResponse processData(DataStorageProcessRequest request) {
         validateRequest(request);
-        
+
+        // Optional: read input data if an input source is provided.
         String inputData = null;
         if (request.getInputSource() != null && !request.getInputSource().isEmpty()) {
             try {
@@ -41,8 +46,18 @@ public class DataStorageProcessAPIImpl implements DataStorageProcessAPI {
             }
         }
 
-        Map<Integer, Integer> productSumResults = computeProductSumNumbers(request.getMinK(), request.getMaxK());
+        // For each k in the range, delegate computation to EngineProcessAPI.
+        Map<Integer, Integer> productSumResults = new HashMap<>();
+        for (int k = request.getMinK(); k <= request.getMaxK(); k++) {
+            // Create an EngineInput for the current k (assuming EngineInput has a constructor that accepts k)
+            EngineInput engineInput = new EngineInput(k);
+            EngineOutput output = engineProcessAPI.compute(engineInput);
+            // Extract the computed product-sum number from the output.
+            // (Assuming EngineOutput has a method getAnswer() returning the computed number.)
+            productSumResults.put(k, output.answer());
+        }
 
+        // Optional: write the results to an output file if an output destination is provided.
         if (request.getOutputDestination() != null && !request.getOutputDestination().isEmpty()) {
             try {
                 FileWriterUtil.writeFile(request.getOutputDestination(), productSumResults.toString());
@@ -57,10 +72,10 @@ public class DataStorageProcessAPIImpl implements DataStorageProcessAPI {
     }
 
     /**
-     * checks the request values:
-     * Request is not null
-     * minK is positive
-     * maxK is greater than or equal to MinK
+     * Validates the incoming request ensuring:
+     * - The request is not null.
+     * - minK is positive.
+     * - maxK is greater than or equal to minK.
      */
     private void validateRequest(DataStorageProcessRequest request) {
         if (request == null) {
@@ -70,69 +85,7 @@ public class DataStorageProcessAPIImpl implements DataStorageProcessAPI {
             throw new IllegalArgumentException("minK must be positive");
         }
         if (request.getMaxK() < request.getMinK()) {
-            throw new IllegalArgumentException("maxK must be greater than or equal to MinK");
-        }
-    }
-
-    /**
-     * Computes minimal product-sum numbers for each k in the range [minK, maxK].
-     * Returns a map where:
-     * Key: k value
-     * Value: minimal product-sum number for that k
-     */
-    private Map<Integer, Integer> computeProductSumNumbers(int minK, int maxK) {
-        Map<Integer, Integer> productSumNumbers = new HashMap<>();
-
-        for (int k = minK; k <= maxK; k++) {
-            int minProductSum = findMinimalProductSum(k);
-            productSumNumbers.put(k, minProductSum);
-        }
-
-        return productSumNumbers;
-    }
-
-    /**
-     * Finds the minimal product-sum number for a specific k value.
-     * A product-sum number must be both the sum and product of exactly k numbers.
-     */
-    private int findMinimalProductSum(int k) {
-        return generateMinimalProductSum(k);
-    }
-
-    /**
-     * Generates the minimal product-sum number for a given k using backtracking.
-     * The algorithm explores possible combinations of numbers that can form both
-     * a sum and product equal to N with exactly k terms.
-     * 
-     * For example, for k=4:
-     * One possible solution is N=8 because:
-     * 8 = 1 + 1 + 2 + 4 (sum)
-     * 8 = 1 × 1 × 2 × 4 (product)
-     */
-    private int generateMinimalProductSum(int k) {
-        Set<Integer> productSumNumbers = new HashSet<>();
-        backtrack(1, 0, 1, k, productSumNumbers);
-        return productSumNumbers.stream().min(Integer::compareTo).orElse(k * 2);
-    }
-
-    /**
-     * Recursive backtracking function to find all possible product-sum numbers.
-     * 
-     * @param sum: Current sum of the numbers
-     * @param count: Current count of numbers used
-     * @param product: Current product of the numbers
-     * @param k: Target number of terms needed
-     * @param results: Set to store valid product-sum numbers found
-     */
-    private void backtrack(int sum, int count, int product, int k, Set<Integer> results) {
-        if (count > 1 && sum == product) {
-            results.add(sum);
-        }
-        if (count >= k || sum > product * 2) {
-            return;
-        }
-        for (int i = 1; i <= product * 2; i++) {
-            backtrack(sum + i, count + 1, product * i, k, results);
+            throw new IllegalArgumentException("maxK must be greater than or equal to minK");
         }
     }
 }
